pyproject.toml
[project]
name = "cocreator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.21"
dependencies = [
    "click>=8.1.8",
    "httpx[socks]>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.0.2",
    "openai>=2.30.0",
    "pillow>=11.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "typer>=0.23.2",
]

[project.scripts]
cocreator = "cocreator.cli:app"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.15.8",
    "ty>=0.0.26",
]
